refactor: move desktop service into app shell
