+++ mpvqc/services/__init__.py
@@ -3,7 +3,6 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from .application_paths import ApplicationPathsService as ApplicationPathsService
-from .desktop import DesktopService as DesktopService
 from .file_startup import FileStartupService as FileStartupService
 from .resource import read_input_conf as read_input_conf
 from .resource import read_mpv_conf as read_mpv_conf
