+++ mpvqc/shell/wiring.py
@@ -14,11 +14,12 @@ def bindings(binder: inject.Binder) -> None:
     import inject
     from PySide6.QtCore import QSettings
 
-    from mpvqc.shell.services import QuitService, ShellSettingsService, VersionCheckerService
+    from mpvqc.shell.services import DesktopService, QuitService, ShellSettingsService, VersionCheckerService
 
     def shell_settings_service() -> ShellSettingsService:
         return ShellSettingsService(inject.instance(QSettings))
 
+    binder.bind_to_constructor(DesktopService, DesktopService)
     binder.bind_to_constructor(QuitService, QuitService)
     binder.bind_to_constructor(ShellSettingsService, shell_settings_service)
     binder.bind_to_constructor(VersionCheckerService, VersionCheckerService)
