+++ pyproject.toml
@@ -251,7 +251,6 @@ files = [
     "mpvqc/resources.py",
     "mpvqc/services/__init__.py",
     "mpvqc/services/application_paths.py",
-    "mpvqc/services/desktop.py",
     "mpvqc/services/file_startup.py",
     "mpvqc/services/resource.py",
     "mpvqc/services/state.py",
@@ -268,6 +267,7 @@ files = [
     "mpvqc/shell/models/dependencies.py",
     "mpvqc/shell/models/shortcuts.py",
     "mpvqc/shell/services/__init__.py",
+    "mpvqc/shell/services/desktop.py",
     "mpvqc/shell/services/quit.py",
     "mpvqc/shell/services/settings.py",
     "mpvqc/shell/services/version_checker.py",
